+ 添加百分比统计

+++ SHCompress/Compress/SHCompress.h
@@ -20,8 +20,8 @@ extern void SHAlloc2(SHData *sh_data, int d);
 extern void SHAlloc3(SHData *sh_data, int d);
 extern void SHFree(SHData *sh_data);
 
-extern int SHBuild2(SHData *sh_data, float **data_set, int count);
-extern int SHBuild3(SHData *sh_data, float **data_set, int count);
+extern float SHBuild2(SHData *sh_data, float **data_set, int count);
+extern float SHBuild3(SHData *sh_data, float **data_set, int count);
 
 extern void SHCompress2(SHData *sh_data, float *source_data, float *compress_data);
 extern void SHUncompress2(SHData *sh_data, float *compress_data, float *source_data);
